Moved some configs
