+++ src/py_mono_tools/goals/deployers.py
@@ -5,8 +5,8 @@ import pathlib
 import subprocess  # nosec B404
 import typing as t
 
-from py_mono_tools.config import cfg, logger
 from py_mono_tools.goals.interface import Deployer, Language
+from py_mono_tools.utils.config import cfg, logger
 
 
 class PoetryDeployer(Deployer):
